feat(widget-relation): string template support
