+++ packages/netlify-cms-lib-widgets/src/index.ts
@@ -1,0 +1,6 @@
+import * as stringTemplate from './stringTemplate';
+
+export const NetlifyCmsLibWidgets = {
+  stringTemplate,
+};
+export { stringTemplate };
