+++ packages/netlify-cms-lib-widgets/webpack.config.js
@@ -1,0 +1,3 @@
+const { getConfig } = require('../../scripts/webpack.js');
+
+module.exports = getConfig();
