+++ packages/netlify-cms-widget-relation/package.json
@@ -30,6 +30,7 @@
     "immutable": "^3.7.6",
     "lodash": "^4.17.11",
     "netlify-cms-ui-default": "^2.6.0",
+    "netlify-cms-lib-widgets": "^1.0.0",
     "prop-types": "^15.7.2",
     "react": "^16.8.4",
     "uuid": "^3.3.2"
